Put custom trait implementations in a separate file
